src/test/java/ingsw/model/cards/publicoc/ShadeVarietyTest.java
package ingsw.model.cards.publicoc;

import com.google.gson.Gson;
import ingsw.model.cards.patterncard.Box;
import ingsw.utilities.GridCreator;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class ShadeVarietyTest {
    ShadeVariety shadeVariety;
    List<List<Box>> grid;

    @BeforeEach
    void setUp() {
        shadeVariety = new ShadeVariety();
        String gridJson = "[\n" +
                "    [{\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 1,\n" +
                "            \"diceColor\": \"BLUE\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 2,\n" +
                "            \"diceColor\": \"GREEN\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 3,\n" +
                "            \"diceColor\": \"YELLOW\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }],\n" +
                "    [{\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 4,\n" +
                "            \"diceColor\": \"PURPLE\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 5,\n" +
                "            \"diceColor\": \"RED\"\n" +
                "        }\n" +
                "    }],\n" +
                "    [{\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 6,\n" +
                "            \"diceColor\": \"GREEN\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }],\n" +
                "    [{\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\",\n" +
                "        \"dice\": {\n" +
                "            \"faceUpValue\": 3,\n" +
                "            \"diceColor\": \"GREEN\"\n" +
                "        }\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }, {\n" +
                "        \"color\": \"BLANK\"\n" +
                "    }]\n" +
                "]";
        grid = new Gson().fromJson(gridJson, GridCreator.GRID_TYPE);
    }

    @Test
    void toStringTest() {
        assertEquals("ShadeVariety", shadeVariety.getName());
        assertEquals("PublicObjCard{'ShadeVariety'}", shadeVariety.toString());
    }

    @Test
    void checkTest() {
        assertEquals(1, shadeVariety.check(grid));
    }

    @Test
    void getScoreTest() {
        assertEquals(5, shadeVariety.getScore(grid));
    }

}